WIP Parallelized tests and small fixes
